submitting all files

+++ src/main/java/com/vehicleRental/app/payloads/CentreDto.java
@@ -14,10 +14,10 @@ import javax.validation.constraints.*;
 public class CentreDto extends BaseEntity {
 
 	@DecimalMin(value = "0.0000001", message = "Latitude should be greater than 0")
-	private long latitude;
+	private double latitude;
 
 	@DecimalMin(value = "0.0000001", message = "Longitude should be greater than 0")
-	private long longitude;
+	private double longitude;
 
 	@NotEmpty(message = "phone cannot be empty")
 	private String phone;
